src/workflow-engine/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
httpx>=0.26.0
asyncpg>=0.29.0
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation-fastapi>=0.43b0
opentelemetry-exporter-otlp-proto-grpc>=1.22.0
langgraph>=0.2.0
langgraph-checkpoint-postgres>=2.0.0
psycopg[binary]>=3.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
websockets>=12.0
pyyaml>=6.0.0
